requirements.txt
Pillow
numpy
opencv_python
scikit_learn
torch
tqdm
visdom
yacs